--- a/TestProjects/VisualEffectGraph/Assets/AllTests/VFXTests/GraphicsTests/SubgraphEventTransmission/animals/cat.png.meta
+++ b/TestProjects/VisualEffectGraph/Assets/AllTests/VFXTests/GraphicsTests/SubgraphEventTransmission/animals/cat.png.meta
@@ -63,7 +63,19 @@ TextureImporter:
     maxTextureSize: 2048
     resizeAlgorithm: 0
     textureFormat: -1
-    textureCompression: 1
+    textureCompression: 0
+    compressionQuality: 50
+    crunchedCompression: 0
+    allowsAlphaSplitting: 0
+    overridden: 0
+    androidETC2FallbackOverride: 0
+    forceMaximumCompressionQuality_BC6H_BC7: 0
+  - serializedVersion: 3
+    buildTarget: Standalone
+    maxTextureSize: 2048
+    resizeAlgorithm: 0
+    textureFormat: -1
+    textureCompression: 0
     compressionQuality: 50
     crunchedCompression: 0
     allowsAlphaSplitting: 0