--- a/com.unity.visualeffectgraph/Editor/Models/Contexts/Implementations/HDRP.meta
+++ b/com.unity.visualeffectgraph/Editor/Models/Contexts/Implementations/HDRP.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2c811597e8869734bbf369237a9bc4d5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: