--- a/com.unity.visualeffectgraph/Editor/Models/Contexts/Implementations/VFXBasicCubeOutput.cs
+++ b/com.unity.visualeffectgraph/Editor/Models/Contexts/Implementations/VFXBasicCubeOutput.cs
@@ -58,5 +58,15 @@ namespace UnityEditor.VFX
         {
             public Texture2D mainTexture = VFXResources.defaultResources.particleTexture;
         }
+        protected override IEnumerable<string> filteredOutSettings
+        {
+            get
+            {
+                foreach (var setting in base.filteredOutSettings)
+                    yield return setting;
+
+                yield return "colorMappingMode";
+            }
+        }
     }
 }