--- a/com.unity.visualeffectgraph/Editor/Models/Contexts/Implementations/VFXLineOutput.cs
+++ b/com.unity.visualeffectgraph/Editor/Models/Contexts/Implementations/VFXLineOutput.cs
@@ -27,6 +27,7 @@ namespace UnityEditor.VFX
                     yield return setting;
 
                 yield return "cullMode";
+                yield return "colorMappingMode";
             }
         }