--- a/com.unity.visualeffectgraph/Editor/Models/Contexts/Implementations/VFXPointOutput.cs
+++ b/com.unity.visualeffectgraph/Editor/Models/Contexts/Implementations/VFXPointOutput.cs
@@ -19,6 +19,7 @@ namespace UnityEditor.VFX
                     yield return setting;
 
                 yield return "cullMode";
+                yield return "colorMappingMode";
             }
         }