--- a/com.unity.visualeffectgraph/Shaders/ParticleHexahedron/Pass.template
+++ b/com.unity.visualeffectgraph/Shaders/ParticleHexahedron/Pass.template
@@ -10,7 +10,7 @@ VFX_VARYING_PS_INPUTS vert(uint id : SV_VertexID, uint instanceID : SV_InstanceI
 {
 	uint index = (id >> 3) + instanceID * 1024;
 	VFX_VARYING_PS_INPUTS o = (VFX_VARYING_PS_INPUTS)0;
-	
+
 	${VFXLoadAttributesOrCull}
 	${VFXProcessBlocks}
 			
@@ -19,15 +19,15 @@ VFX_VARYING_PS_INPUTS vert(uint id : SV_VertexID, uint instanceID : SV_InstanceI
 	
 	float3 offsets = (float3)0;
 	offsets.x = float(id & 1);
-	offsets.y = float((id & 3) >> 1);
-	offsets.z = float((id & 7) >> 2);
+	offsets.y = float((id & 2) >> 1);
+	offsets.z = float((id & 6) >> 2);
 	offsets -= 0.5f;
 	
 	${VFXLoadSize}
 	
 	float3x3 rot = GetEulerMatrix(radians(float3(angleX,angleY,angleZ)));
 	float4x4 elementToVFX = GetElementToVFXMatrix(axisX,axisY,axisZ,rot,float3(pivotX,pivotY,pivotZ),size3,position);
-	
+
 	float3 vPos = mul(elementToVFX,float4(offsets,1.0f)).xyz;
 	float3 vPosWS = TransformPositionVFXToWorld(vPos);
 	
@@ -49,17 +49,33 @@ VFX_VARYING_PS_INPUTS vert(uint id : SV_VertexID, uint instanceID : SV_InstanceI
 	o.VFX_VARYING_ROTY = rot[1];
 #endif
 	o.VFX_VARYING_POSCS = VFXTransformPositionWorldToClip(vPosWS);
-
+	
 	${VFXVertexCommonProcess}
+	
+	
+	#if USE_UV_SCALE_BIAS && defined(VFX_VARYING_UV_SCALE) && defined(VFX_VARYING_UV_BIAS)
+	o.VFX_VARYING_UV_SCALE = uvScale;
+	o.VFX_VARYING_UV_BIAS = uvBias;
+	#endif
 	
 	#if USE_FLIPBOOK
 	${VFXLoadParameter:{flipBookSize}}
 	${VFXLoadParameter:{invFlipBookSize}}
+	#ifdef VFX_VARYING_TEXINDEX
 	${VFXLoadParameter:{texIndex}}
+	#endif
 	
+	#if defined(VFX_VARYING_FLIPBOOKSIZE) && defined(VFX_VARYING_INVFLIPBOOKSIZE)
 	o.VFX_VARYING_FLIPBOOKSIZE = flipBookSize;
 	o.VFX_VARYING_INVFLIPBOOKSIZE = invFlipBookSize;
+	#endif
+	#ifdef VFX_VARYING_TEXINDEX
 	o.VFX_VARYING_TEXINDEX = texIndex;
+	#if USE_FLIPBOOK_MOTIONVECTORS && defined(VFX_VARYING_MOTIONVECTORSCALE)
+	${VFXLoadParameter:{motionVectorScale}}
+	o.VFX_VARYING_MOTIONVECTORSCALE = motionVectorScale * invFlipBookSize;
+	#endif
+	#endif
 	#endif
 
 	${VFXVertexAdditionalProcess}