--- a/com.unity.visualeffectgraph/Shaders/ParticleHexahedron/PassBasicForward.template
+++ b/com.unity.visualeffectgraph/Shaders/ParticleHexahedron/PassBasicForward.template
@@ -19,15 +19,34 @@ Pass
 		// y: alpha threshold
 		nointerpolation float2 builtInInterpolants : TEXCOORD1;
 		#endif
-		nointerpolation uint faceID : TEXCOORD2;
+		#if USE_FLIPBOOK_MOTIONVECTORS
+		// x: motion vectors scale X
+		// y: motion vectors scale Y
+		nointerpolation float2 builtInInterpolants2 : TEXCOORD2;
+		#endif
+		nointerpolation uint faceID : TEXCOORD3;
 		#if VFX_NEEDS_POSWS_INTERPOLATOR
-		float3 posWS : TEXCOORD3;
+		float3 posWS : TEXCOORD4;
+		#endif
+		#if USE_UV_SCALE_BIAS
+		nointerpolation float4 scaleBias : TEXCOORD5;
+		#endif
+		#if USE_FLIPBOOK
+		nointerpolation float4 flipBookSize : TEXCOORD6;
+		nointerpolation float texIndex : TEXCOORD7;
+		#endif
+		#if WRITE_MOTION_VECTOR_IN_FORWARD
+		float4 cPosPrevious : TEXCOORD4;
+		float4 cPosNonJiterred : TEXCOORD5;
 		#endif
 	};
 	
 	struct ps_output
 	{
 		float4 color : SV_Target0;
+#if WRITE_MOTION_VECTOR_IN_FORWARD
+		float4 outMotionVector : SV_Target1;
+#endif
 	};
 
 #define VFX_VARYING_PS_INPUTS ps_input
@@ -41,7 +60,24 @@ Pass
 #if VFX_NEEDS_POSWS_INTERPOLATOR
 #define VFX_VARYING_POSWS posWS
 #endif
+#if USE_UV_SCALE_BIAS
+#define VFX_VARYING_UV_SCALE scaleBias.xy
+#define VFX_VARYING_UV_BIAS scaleBias.zw
+#endif
+#if USE_FLIPBOOK
+#define VFX_VARYING_TEXINDEX texIndex
+#define VFX_VARYING_FLIPBOOKSIZE flipBookSize.xy
+#define VFX_VARYING_INVFLIPBOOKSIZE flipBookSize.zw
+#endif
+#if USE_FLIPBOOK_MOTIONVECTORS
+#define VFX_VARYING_MOTIONVECTORSCALE builtInInterpolants2.xy	
+#endif
 								
+#if WRITE_MOTION_VECTOR_IN_FORWARD
+#define VFX_VARYING_VELOCITY_CPOS cPosNonJiterred
+#define VFX_VARYING_VELOCITY_CPOS_PREVIOUS cPosPrevious
+#endif
+
 	${VFXInclude("Shaders/ParticleHexahedron/Pass.template")}	
 		
 	#pragma fragment frag
@@ -49,8 +85,8 @@ Pass
 	{
 		ps_output o = (ps_output)0;
 		VFXTransformPSInputs(i);
-		
 		o.color = VFXGetFragmentColor(i);
+		
 		o.color *= VFXGetTextureColorWithProceduralUV(VFX_SAMPLER(mainTexture),i,VFXCubeGetUV(i));
 		o.color = VFXApplyFog(o.color,i);
 		VFXClipFragmentColor(o.color.a,i);