… TestProjects/VisualEffectGraph/Assets/AllTests/VFXTests/GraphicsTests/SubgraphEventTransmission/animals/cat.png