* Update HDCameraUI.Skin.cs

* Improve Variant Provider interface (#73)

* Add overridable ComputeVariant function for variant provider

* Update changelog

* Improve AttributeFromCurve Variant provider & Remove Curl Cellular from listing

* Remove ".XYZ" for library naming

* Prevent to build a player if no hdrp asset found (#2729)

* Vector4 Field & Animated Parameter (#56)

* Clean approach :
- Use MultiPropertyField instad of Vector4Field
- Use Begin/End property instead of IsAnimated

* Fix inspector while resizing

* Update changelog + rebuild vfx

* Hdrp fix open gl error message (#2733)

* Fix OpenGL build error not displayed in the editor

* Removed useless comments

* Camera-relative support for stereo (#2719)

* WIP - attempt to fix camera-relative directional shadows with stereo

* Add ApplyCameraRelativeStereoOffset() to avoid dup code

* Fix shadowmaps with camera-relative

* shader code cleanup and fixes

* refactor C# code for stereo camera-relative

* add entry to changelog

* move calls to StereoCameraRelativeEyeToCenter to higher level to avoid applying offset twice

* HD/Post-processing fixes (#2732)

* Cleaned up & simplified previous PR on pre-exposure control

* Added guard bands to bloom to deal with RTHandle resizing

Hack, but better than leaking stuff from the top & left of the screen

* Motion blur for PP v3 (#2727)

* Manual rebase of wip-motion-blur (modified files)

* Resubmit editor

* Resubmit component

* Resubmit shaders

* Fixup resource asset

* Set intensity to 0 by default

* Revert the no motion set to 2.0

* Use Wave Intrinsics tile generation (2x improvement on PS4, 1.5x on Scorpio)

* Remove int division

* Better threshold for vel prep

* Revert (For now) the thresh check on vel prep stage

* Refactor parallel reduction code

* Getting rid of merged prep and tile

* Slight optimization to the wave intrinsic version of the tile generation

* [Testing - probably going to revert] Try scalar load for tile max velocity

* Revert "[Testing - probably going to revert] Try scalar load for tile max velocity"

This reverts commit 844778ac4935d1a5bd3842636b6b468d1c21892f.

* [Testing - probably going to revert] Try scalar load for tile max velocity

This reverts commit 3dafec82c914d555536ce13c4e1c5ca60fc715dd.

* Revert "[Testing - probably going to revert] Try scalar load for tile max velocity"

This reverts commit 0416e62182acdb143d380bbf50f793aba2322fb9.

* Some messing around with debug vis and weights (nothing useful really, most will be reverted)

* Optimize sample count

Minimum of 2 samples, but we clamp the number of sample so that we'll never have samples that span less than half a pixel.

* Fix few warnings

* Reset history flag

* Clamp tile coordinates when computing neighbourhood max

* Change early out clause in vel prepping

* Change blur group size from 8x8 to 16x16. *Much* better (reduce cost by 1/3 in some cases) on Scorpio, slightly better or same on PS4

* PACKING is now not an option anymore, but always the case

* Optimize sample count is not an option but always on now

* Remove f from float numbers

* Move compute velocity and scale rotation by intensity as well

* Add clamp of camera rotation and move a bit of things from common to prep

* Use texture macros

* Make explicit group size variable

* Handle "Force no motion"

* Better jitter for tiles and update tooltip

This reverts commit 95889accd337e93175fa2f2f93c0e318dfbf6db6.

* Fixup RTHandle issues when switching to different sized RT

* Update UI,  update sample count debug mode and a bit of cleanup

* Fix tab issue

* Convert tabs to spaces

* Comment debug symbols pragma

* Better name for sample count debug view

* Comment on the double direction sampling

* Comment on why we set velocity to 2 for "force no motion"

* Update MotionBlur.cs

* Update MotionBlur.cs

* Fix  Metal API not allowed by HDRP (#2736)

* Add enable field to motion blur editor (#2737)

* turn off post process for stereo (#2735)

* HDRP emission node, unit intensity and pre-exposure control (#2722)

* Add Emission Node and pre-exposure control for emission

* Removed emission pre-expose percent for terrain

* Added emission fields for layered lit and layered lit tesselation

* Update CHANGELOG.md

* Added Emission Node doc, and unlit emissive UI

* Use exp2 instead of PositivePow

* Renamed pre-exposure weight to Exposure Weight

* Removed EmissionColorHDR and moved GetExposureMultiplier to

* Add emissive intensity fields for lit tesselation shader

* Better parity for sequential operator (#78)

* Better parity for sequential operator, Unify Sequential3D with count as first entries too
=> TODO, add orientable Circle Type & Box Mode

* Add #pragma editor_sync_compilation to various editor and first time shader

* Update HDRenderPipelineResources.asset

* Change cascade shadow map error message (#2743)

* Fix for case of Template for LINUX compatibility

* Optimized uberfx & final pass (#2749)

Xbox One X (4K):
- Uber without bloom and/or dirt: 540µs -> 368µs (32%)
- Finalpass: 587µs -> 410µs (30%)

* Add tests for motion blur (disabled by default) (#2748)

* Fix planar reflection issue after merge of multiviewport (#2752)

* update reference screenshots (#2753)

* Fix planar reflection issue after merge of multiviewport

* Update reference screenshots

* fixed issue with PBR master node not using correct geometric normal

* Fix issue with mircoshadow and specular occlusion when LightLayers is enabled (#2757)

* fix microshadowing with specular occlusoin and lightlayers

* Update SimpleLit.hlsl

* Added color grading curves (YRGB + secondaries) (#2759)

* HDRP shader graph HD render queues (#2750)

* Added HDMaterialTags to generate shader tags in shader graphs

* Added an error message for objects that have refraction and BeforeRefraction render pass

Plus added a renderpass combobox for unlit shader graph node

* Remove useless comment

* Fix shader graph render pass update when switching opaque/transparent

* Fix transparent render queue generation

* Fixed HDRP shader compilation

* Fixed scaleY in ConnectTarget (was broken when size != 1.0) (#79)

* Fixed scaleY in ConnectTarget (was broken when size != 1.0)

* Added changes to changelog

* Fixing the Lit warnings due to Improved moment shadows (#2762)

* Material validator hdrp (#2738)

* Material validator using Debug Display for HDRP

* Update changlog

Update changelog to include material validation

* Update based on PR feedback

* Update code with PR feedback - Rename validation method

* Update CHANGELOG.md

* Update CHANGELOG.md

* Fix compilation issue + rename mode to diffuseColor / SpecularColor

* Update string in UI

* More renaming

* Fix typo

* Disable Motion blur via intensity instead of checkbox  (#2765)

* Disable mb via intensity

* Ooopsie

* Update for 32 threads in wavefronts on Switch. (#2724)

* Update for 32 threads in wavefronts on Switch.

* removed SORTLIST() and unecessary GroupMemoryBarrierWithGroupSync() to get parity with PS4/XB1.

* Moved thread group size to 32 threads on Switch (minor speed up from 0.66ms to 0.55ms).

* Encapsulated optimal thread group size into a #define.
PS4, XB1 and Switch can take advantage of the optimal group size.

* Removed obsolete comment (the file is not autogenerated anymore).

* Removed debug mode as it causes compiler error on Switch.

* Updated comments.

* Moved THREAD_GROUP_OPTIMAL_SIZE into platform headers and renamed to PLATFORM_THREAD_GROUP_OPTIMAL_SIZE.

* Missing updated file from previous commit.

* Added comments.

* Use GraphicsFormat instead of RenderTextureFormat  (#2761)

* Move to GraphicsFormat

* Fix bug with decals

* Use graphics format instead of format in the pool recycler (#2767)

* Fix RenderQueue update when changing surface type on unlit HD master node (#2768)

* Fixed serialization of SurfaceThickness setting in Collider Depth

* Fix legacy shader build

* Update depth and color node tests for HDRP (#2771)

* Vfx/fix/ui fixes (#81)

* Fix exception when a port having edge is removed.

* Refactor ContextUI RefreshContext

* Fix for block edges not disappearing.

* fix for RefreshContext refactor

* fix slider to change value directly on click.

* Fix index when adding new block.

* fix toobar item alignement.

* No longer restore global selection when window is focused( because we receive a focus event even when only one of the docked panel in our window is focused ).

* New iteration on gizmo : make it compatible with new property storage ( property stored only if overridden ) . Fix local/world on component gizmo once again.

* Gizmos : automatically add new property if the gizmo is modified and the value of the gizmo was not overridden yet.

* Unprepare the visual effect editor gizmos when the state of an overridden property has changed.

* Blackboard now shows level two and more of subproperties (for example ArcCircle.circle.center / .radius) if expanded

* Fix location of new context when dropping a flow edge over the graph [FB:1117090]

* [FB:1117127]Opening the Target Component window by clicking on component "edit" updates the state of the toggle "Target GameObject"

* Remove second "Visual Effects" menu[1117105]

* Put Visual Effect graph into preference category.

* Fix for stickynote sliding when dragging a group it is in.

* Fix case where view can't be found in anchor AttachToPanelEvent

* make sure input and output slot are at the same height

* hide divider on contexts when there is no settings.

* Fix block selection border margins.

* Hide settings dvider if the node has no settings.

* Switch to new notification API for scene view in slotcontainereditor.

* Rotate scale gizmo.

* Added Depth sampling modes to the scene depth node (#2740)

* Added Depth sampling modes to the scene depth node

* Fix eye depth mode description

* Fixed typos

* Update CHANGELOG.md

* Fix warning an issue with scenecolor node in hdrp (#2774)

* Fix warning an issue with scenecolor node in hdrp

* Delete InitTestScene636834054956858632.unity.meta

delete wrong submited files

* Delete InitTestScene636834054956858632.unity

delete wrong submited files

* revert undesired change

* fix for merge

* Hdrp/fix shadergraph hdrenderqueue migration (#2776)

* Fix renderingPass creation along data from prior its existance

* Fix renderingPass creation along data from prior its existance (Unlit)

* Fix After PostProcessTransparent generated queue text to be base on Transparent

* [HDRP] Upgrade FrameSettings (#2766)

* Introduce BitArray utility class

* Introduce FrameSettingsAttribute to help inspector and DebugMenu display

* Add unit test on BitArray

* Refactor FrameSettings and merge LightLoopSettings

* Add data migration

* Rename and update OverridableFrameSettingsArea UI utility

* Introduce FrameSettingsHistory for DebugMenu debugging agregatted values

* Fix FrameSettings usage

* Fix tooltips of Camera

* Fix former unit tests referencing FrameSettings

* Update UI

* Update DebuggerDisplays

* Fix alignment issue in PlanarReflectionProbe and ReflectionProbe inspectors

* Easing add of new FrameSettings in UI part based on FrameSettingField attribute

* Fix Default non bool FrameSettings

* Fix FrameSettings ordering

* Fix auto numbering of FrameSettingsFieldAttribute

* Fix FrameSettingsHistory and do history field UI

* Add DebugMenu generation on the same principe than inspector

* Add history in DebugMenu

* Clean history activation

* Speed up FrameSettingsHistory enabled

* Cleaning

* QuickFix issue due to PreviewCamera becoming of type Game for no reason.

* Add auto fixing broken serialized resources for HDRP.

* Update reflection method call for Preview of camera to reflect updated trunk

* Add support for multiselection

* Missing migration of recently added ExposureControl

* Update reflection method call for Preview of PlanarReflectionProbe to reflect updated trunk

* Update changelog

* Fix issue when adding-removing Scene View tab

* Fix typos

* Update FrameSettings.Migration.cs

* Prenvent very old probes to crash at migration when they have no LightLoopSettings

* Prevent SceneCameras to feed several times debug menu when HDRenderPipeline is reconstructed

For instance, this occures when we launch graphics tests

* Update MiniProfiler and debug menu which was broken

* Revert "Update MiniProfiler and debug menu which was broken"

This reverts commit 434b5a14f52075cf259f6f4958525a57ff27a390.

* Update FrameSettings.cs

* Update 2301_Shadow_Mask.unity

* Primary Visibility and Area Shadow (plus some additional raytracing features) (#2770)

* Adding a Raytracing render queue for both transparent and opaque objects
Opaque Objects that are in the raytracing queue are rendered in the prepass and not in the deferred/forward passes
Adding the primary visibility mode in the raytracing environement
Correcting a label in the rt_env_inspector
Adding a Debug Mode for raytracing area shadow
Adding an equivalent of the ssr lit brdf evaluation
Adding a forward raytracing pass for the lit and the unlit
Supporting Ior for the lit.shader
Implementing a HDRaytracingRenderer that handles primary visibility code
Adding resources for the raytracing primary ray
Not Pushing a light to the Light cluster if null (has been deleted)
Enabling Automatic raytracing scene invalidation on hierarchy change in the editor
Using the global command buffer parameter instead of the raytracing one when required for closest hit
Adding a texture for the analytic value (required for shadow filtering)
Adding default anyhit shaders to ambient occlusion and area shadow raytracing shaders
Adding the remainingDepth to the rayintersection structure
Adding an input reflection and transmission value for the Raytracing Lightloop and adding them to the evaluation
Adding the light dimension to the spherical rectangle
Correcting the bias usage for Raytraced AO and Reflection

* Refactoring and stabillization changes for the raytracing code

* Review corrections

* Order in args check for raytracingrenderer

* Delete 9601_MotionBlur.meta

* Update reference screenshots

* Avoid issue with modulo (save actually some computation recycling result of unsigned division anyway) (#83)

* remove unsupported FormerlySerializedAs in ShaderGraph serialisation (#2779)

* Hdrp/fix framesettings formerly nullable tests (#2781)

* Remove test: FrameSettings cannot be null anymore.

* Change test: FrameSettings cannot be null anymore. But wrong rendertexture type.

* Update 8205 test

* Add editor test to cover fixed delta time (using an debug spawner state) (#84)

* Update graphic test (scene + screenshots)

* Update HDCubemapInspector.cs

* Vfx/fix/add prefab editor test 1117103 (#80)

* Add new test with expected failure (commented in l.171 & 172) : Covers simple prefab override from fogbugz 1117103

* Add other test to cover future possible issue due to refactor of serialization cycle in VisualEffect component

* Minor : change exposed name

* Hdrp master reflection fix (#2785)

Fix of debug mode with reflection probe

* Hdrp/fix framesettings msaa and exposurecontrol (#2786)

* Add negativeDependency in FrameSettings

* Make MSAA's frame settings more user friendly

* Fixed sky exposure (#2778)

Fixed sky exposure

* Update CHANGELOG.md

* Hdrp/new ambient mode (#2630)

* - Added AmbientMode enum
- Renamed Baking Sky to Static Lighting Sky

* Implemented static and dynamic sky ambient probe

* Optimized version of AmbientProbeConvolution

* Removed manual serialization from static lighting sky

* Increased the number of sample to 256 (optimized with LDS on PS4/XBone)

* Fixed ambient probe flickering when a preview is displayed

* Changelog + Comment fixes

* Updated Test data according to static lighting changes

* - Fixed deprecated calls
- Fixed static lighting when no baking sky is present
- Fixed deserialization of former BakingSky component.

* Fixes/comments from PR feedback

* Removed f from float values in ambient convolution shader.

* Refactor to make realtime GI work with ambient static/dynamic modes

* Fixed issue with missing skybox material when running tests.

* Updated screenshots

* Temporarily disabled realtime GI part of test 2102

* Update Screenshots after merge

* Update scenes and screenshots

* Rebuild vfx

* Fix shader compilation issue  (#2793)

* Fix shader compilation issue

* remove old def on PS4

* Hdrp/fix framesettings probe overrides and tests (#2794)

* Make DebuggerDisplay easier to read between FrameSettings and FrameSettingsOverideMask

* Fix FrameSettings override for ReflectionProbes

* Fix FrameSettings Equals

* Fix RandomUtilities.RandomFloat and add RandomEnumValue

* Add unit tests for framesettings aggregation

* Ammend CameraSettingsUtilitiesTests.ApplySettings test with FrameSettings check

* Fix missing ShaderLitMode in FrameSettingsMigration

* Added migration test for FrameSettings

* update vfx test screenshot for master

* Update skyManager with  UnityEditor.Lightmapping api changed

* Hdrp fix double sided (#2798)

* Fix the flip for the Lit shader

* Move flip normal handling to the resolve point

* Remove bitangent flips

* Fix the double sided 'None' mode

* Unify code of Lit and SG

* update init values

* comment //#pragma enable_d3d11_debug_symbols

* Update reference screenshots

* FrameSettings SSR & Raytracing Fixes (#2801)

* Corrections due to merge issues

* SSR FrameSetting condition correction

* Update 2103 test

* Ambient occlusion raytracing hit distance (#2802)

Correcting type in the raytracing enable flag

* Add inverse lerp node (#85)

* Corretions for the render queue field in shader graph hdrp shaders (#2787)

* Corretions for the render queue field in shader graph hdrp shaders

* Only display working rendering pass for now for surfacetype

* Some cleaning

* Fixed issues with realtime probe resolution (#2806)

* - Fixed an issue where probe target texture wasn't resized properly when changing HDRP asset.
- Fixed rendering viewport for probe rendering

* Updated obsolete screenshot (exposure issue)

* Update changelog

* Rebuild vfx

* Update bad reference image

* Update VFX Test project settings

* Revert "Update VFX Test project settings"

This reverts commit 9d4e2ec231870b80b35b6afa4be42da0ac4449df.

* Revert unwanted changes

* Vfx/feature/expose light probe usage (#77)

* First implementation

* Encapsulate all VFX renderer related inspector stuff in RendererEditor class

* Add proxy override and probe anchor settings

* Update changelog

* Fix typo

* fix for override check width.

* fixed slider field focus

* disable blackboard add button is no visualeffect asset is displayed.

* Multiple fixes needed when more than one VisualEffect is selected (FB #1110747) (#90)

* Multiple fixes needed when more than one VisualEffect is selected (FB #1110747)

* reseed button give each selected component a different value.

* Refactored VFManagerEditor to cache properties. Guess hen reset has been selected to reset shaders and path to sane value.

* Fix missing inspector for VisualEffectActivationClip

* Use system title first line ( if any ) in Aset inspector output list.

* fix exception when restoring a removed context class.

* Update VFX

…